% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.adjacency.R
\name{convert.adjacency}
\alias{convert.adjacency}
\title{convert.adjacency -  A function where you input some kind of adjacency object
and you out the object class of your choice.}
\usage{
convert.adjacency(A, out = "matrix", df = NULL)
}
\arguments{
\item{A}{The adjacency object to convert of class type
c("matrix", "dgCMatrix", "edgelist", "igraph", "network").}

\item{out}{The form of the adjacency object to output. Options are
c("matrix", "Matrix", "edgelist", "igraph", "network").}

\item{df}{An optional data frame of node attributes. Only potentially
important if out="igraph" so the outputted igraph object has node attributes.}
}
\value{
An object of class=out:

\code{A} The adjacency matrix.

\code{df} A data frame of node attributes. Mostly applicable if input
is an "igraph" object with node attributes.
}
\description{
Will fail if the object is too large to store in memory as class "matrix"
and the original object is a network or out="matrix". (no method
for coercing a "network" object directly into a dgCMatrix)
}
\details{
Additionally, you cannot input a row-normalized matrix and output an igraph
or edgelist, it will not be correct: try running row.norm(A, reverse=TRUE).
}
