% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgenet.R
\docType{methods}
\name{edgenet,matrix,matrix-method}
\alias{edgenet,matrix,matrix-method}
\title{@rdname edgenet-methods}
\usage{
\S4method{edgenet}{matrix,matrix}(X, Y, G.X = NULL, G.Y = NULL,
  lambda = 1, psigx = 1, psigy = 1, thresh = 1e-05,
  maxit = 1e+05, learning.rate = 0.01, family = gaussian)
}
\description{
@rdname edgenet-methods
}
