% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgenet.R
\docType{methods}
\name{edgenet}
\alias{edgenet}
\title{Fit a graph-regularized linear regression model using
 edge-based regularization.}
\usage{
edgenet(X, Y, G.X = NULL, G.Y = NULL, lambda = 1, psigx = 1,
  psigy = 1, thresh = 1e-05, maxit = 1e+05, family = gaussian)
}
\arguments{
\item{X}{input matrix, of dimension (\code{n} x \code{p})
where \code{n} is the number of observations and \code{p} is the number
of covariables. Each row is an observation vector.}

\item{Y}{output matrix, of dimension (\code{n} x \code{q})
where \code{n} is the number of observations and \code{q} is the number
of response variables. Each row is an observation vector.}

\item{G.X}{non-negativ affinity matrix for \code{n}, of dimensions
(\code{p} x \code{p}) where \code{p} is the number of covariables \code{X}}

\item{G.Y}{non-negativ affinity matrix for \code{n}, of dimensions
(\code{q} x \code{q}) where \code{q} is the number of responses \code{Y}}

\item{lambda}{\code{numerical} shrinkage parameter for LASSO.}

\item{psigx}{\code{numerical} shrinkage parameter for graph-regularization
of \code{G.X}}

\item{psigy}{\code{numerical} shrinkage parameter for graph-regularization
of \code{G.Y}}

\item{thresh}{\code{numerical} threshold for optimizer}

\item{maxit}{maximum number of iterations for optimizer
(\code{integer})}

\item{family}{family of response, e.g. \emph{gaussian} or \emph{binomial}}
}
\value{
An object of class \code{edgenet}
\item{coefficients }{ the estimated (\code{p} x \code{q})-dimensional
 coefficient matrix B.hat}
\item{intercept }{ the estimated (\code{q} x \code{1})-dimensional
 vector of intercepts}
\item{call }{ the call that produced the object}
\item{family }{ the family of the response}
}
\description{
Fit a graph-regularized linear regression model using
 edge-penalization. The coefficients are computed using graph-prior
 knowledge in the form of one/two affinity matrices. Graph-regularization is
 an extension to previously introduced regularization techniques,
 such as the LASSO. See the vignette for details on the objective function of
 the model: \href{../doc/edgenet.html}{\code{vignette("edgenet", package="netReg")}}
}
\examples{
X <- matrix(rnorm(100*10), 100, 10)
b <- rnorm(10)
G.X <- matrix(rpois(100,1), 10)
G.X <- t(G.X) + G.X
diag(G.X) <- 0

# fit a Gaussian model
Y <- X \%*\% b + rnorm(100)
fit <- edgenet(X=X, Y=Y, G.X=G.X, family=gaussian)
}
\references{
Dirmeier, Simon and Fuchs, Christiane and Mueller, Nikola S and Theis,
 Fabian J (2018),
 netReg: Network-regularized linear models for biological association
 studies. \cr
 \emph{Bioinformatics}\cr \cr
 Friedman J., Hastie T., Hoefling H. and Tibshirani R. (2007),
 Pathwise coordinate optimization.\cr
 \emph{The Annals of Applied Statistics}\cr \cr
 Friedman J., Hastie T. and Tibshirani R. (2010),
 Regularization Paths for Generalized Linear Models via
 Coordinate Descent.\cr \emph{Journal of Statistical Software}\cr \cr
 Fu W. J. (1998),  Penalized Regression: The Bridge Versus the Lasso.\cr
 \emph{Journal of Computational and Graphical Statistics}\cr \cr
 Cheng W. and Wang W. (2014), Graph-regularized dual Lasso for robust
 eQTL mapping.\cr
 \emph{Bioinformatics}
}
