% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\docType{methods}
\name{gaussian}
\alias{gaussian}
\alias{binomial}
\alias{poisson}
\title{Family objects for models}
\usage{
gaussian(link = c("identity"))

binomial(link = c("logit", "probit", "log"))

poisson(link = c("log"))
}
\arguments{
\item{link}{name of the link function}
}
\value{
An object of class \code{netReg.family}
 \item{family }{ name of the family}
 \item{link }{ name of the link function}
 \item{linkinv }{ inverse link function}
 \item{loss }{ loss function}
}
\description{
Family objects provide a convenient way to specify the details
 of the models used by \code{netReg}.
 See also \code{\link[stats:family]{stats::family}} for more details.
}
