% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lasso.R
\docType{methods}
\name{group.lasso}
\alias{group.lasso}
\alias{group.lasso,matrix,numeric-method}
\alias{group.lasso,matrix,matrix-method}
\title{Fit a linear regression model using a group lasso penalty}
\usage{
group.lasso(
  X,
  Y,
  grps = NULL,
  lambda = 1,
  thresh = 1e-05,
  maxit = 1e+05,
  learning.rate = 0.01,
  family = gaussian
)

\S4method{group.lasso}{matrix,numeric}(
  X,
  Y,
  grps = NULL,
  lambda = 1,
  thresh = 1e-05,
  maxit = 1e+05,
  learning.rate = 0.01,
  family = gaussian
)

\S4method{group.lasso}{matrix,matrix}(
  X,
  Y,
  grps = NULL,
  lambda = 1,
  thresh = 1e-05,
  maxit = 1e+05,
  learning.rate = 0.01,
  family = gaussian
)
}
\arguments{
\item{X}{input matrix, of dimension (\code{n} x \code{p})
where \code{n} is the number of observations and \code{p} is the number
of covariables. Each row is an observation vector.}

\item{Y}{output matrix, of dimension (\code{n} x \code{q})
where \code{n} is the number of observations and \code{q} is the number
of response variables. Each row is an observation vector.}

\item{grps}{vector of integers or \code{NA_integer_} of length \code{p}
that encodes the grouping of variables, e.g., \code{c(1, 1, 2, 2, NA)}}

\item{lambda}{\code{numerical} shrinkage parameter}

\item{thresh}{\code{numerical} threshold for optimizer}

\item{maxit}{maximum number of iterations for optimizer
(\code{integer})}

\item{learning.rate}{step size for Adam optimizer (\code{numerical})}

\item{family}{family of response, e.g., \emph{gaussian} or \emph{binomial}}
}
\value{
An object of class \code{edgenet}
\item{beta }{ the estimated (\code{p} x \code{q})-dimensional
 coefficient matrix B.hat}
\item{alpha }{ the estimated (\code{q} x \code{1})-dimensional
 vector of intercepts}
\item{parameters }{ regularization parameters}
\item{lambda }{ regularization parameter lambda)}
\item{family }{ a description of the error distribution and link function
   to be used. Can be a \code{\link[netReg:family]{netReg::family}} function or a character string
   naming a family function, e.g. \code{gaussian} or "gaussian".}
\item{call }{ the call that produced the object}
}
\description{
Fit a linear regression model the group LASSO penalty.
}
\examples{
X <- matrix(rnorm(100 * 10), 100, 5)
b <- rnorm(5)
grps <- c(NA_integer_, 1L, 1L, 2L, 2L)

# estimate the parameters of a Gaussian model
Y <- X \%*\% b + rnorm(100)
fit <- group.lasso(X = X, Y = Y, grps = grps, family = gaussian, maxit = 10)

}
\references{
Yuan, Ming and Lin, Yi (2006),
 Model selection and estimation in regression with grouped variables. \cr
 \emph{Journal of the Royal Statistical Society: Series B}\cr \cr
 Meier, Lukas and Van De Geer, Sara and Bühlmann, Peter (2008),
 The group lasso for logistic regression. \cr
 \emph{Journal of the Royal Statistical Society: Series B}
}
