% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hand.R
\name{hand}
\alias{hand}
\title{hand - A function to perform the Homophily Adjusted Network Disturbances Model}
\usage{
hand(
  formula,
  data,
  network,
  method,
  beta_prior = list(mean = 0, sigma = 1),
  s2_prior = list(shape = 2, scale = 0.5),
  rho_prior = list(mean = 0.36, sd = 0.7^2),
  ...
)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced
to that class): a symbolic description of the model to be fitted.}

\item{data}{an optional data frame, list or environment (or object coercible
by as.data.frame to a data frame) containing the variables in the model.
If not found in data, the variables are taken from environment(formula),
typically the environment from which nam is called. You can also input an
igraph object and the dataset and adjacency matrix will be extracted for
you (in which case the network argument will be ignored).}

\item{network}{An adjacency matrix describing your network. Can be of the class:
c("matrix", "dgCMatrix", "igraph", "network") or an edgelist.}

\item{method}{A character indicating the method to use for simulation.
Options vary by model but all potential options are c("normal",
"greta", "stan", "lnam"). Run ?methods.nam() for model-specific designations.}

\item{beta_prior}{The prior for the beta vector of the model. Requires list
or function inputs. Default is a multivariate normal with mean=0 and sigma=1
If you supply a list, you must have one element named 'mean' which is the
mean of the multivariate normal and one named 'sigma' which is the
variance-covariance matrix. If mean is a single integer,
that mean will be used for every element in the mean vector and if sigma is
a single integer, it will multiply that integer by the identity matrix.
Otherwise, both mean and sigma will be what you specify them to be. Or the
user can specify their own prior function with which to input. If
method="greta", the user-specified function must be a function from the
greta package.}

\item{s2_prior}{Prior for the variance parameter. Requires list or function
inputs. Defaults to a gamma(2,0.5). If Input is a list, the list must have
two elements named 'shape' and 'scale' to represent the shape and scale of
the gamma distribution respectively. Or the user can specify their own prior
function with which to input. If method="greta", the user-specified function
must be a function from the greta package.}

\item{rho_prior}{The prior for the correlation of the adjacency matrix of
the model. Requires list or function inputs. Default is a N(0.36,0.49)
(Dittrich et al. 2017). If input is a list, the list must have two elements
named 'mean' and 'sd' which are the mean and standard deviation of the normal
distribution respectively. Or the user can specify their own prior
function with which to input. If method="greta", the user-specified function
must be a function from the greta package. If method="stan" only list input
will be accepted.}

\item{...}{Additional arguments to be passed into the optim() function for
when method="normal".}
}
\value{
A list with two classes, the first being "nam" and the second is the
class of the method that contains values:

\code{parameters} A list containing the parameters of the chosen method.

\code{A} The adjacency matrix. Will be row-normalized if rownorm=TRUE.

\code{loglik} The log of the likelihood of the model.

\code{AIC} The Akaike Information Criterion of the model.

\code{BIC} The Bayesian Information Criterion of the model.
}
\description{
hand - A function to perform the Homophily Adjusted Network Disturbances Model
}
