% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv.R
\name{IV}
\alias{IV}
\title{IV - A function to perform the Kelejian and Prucha (SARAR) Model}
\usage{
IV(
  formula,
  data = NULL,
  network,
  method,
  beta_prior = list(mean = 0, sigma = 1),
  s2_prior = list(shape = 2, scale = 0.5),
  rho_prior = list(mean = 0.36, sd = 0.7^2),
  lambda_prior = list(mean = 0.36, sd = 0.7^2),
  rownorm = TRUE,
  ...
)
}
\arguments{
\item{formula}{symbolic description of the model to be fitted.}

\item{data}{optional data frame or environment.}

\item{network}{Adjacency matrix.}

\item{method}{Estimation method: c("normal", "IS", "stan").}

\item{beta_prior}{List or function for beta vector.}

\item{s2_prior}{List or function for variance.}

\item{rho_prior}{List or function for the outcome lag correlation.}

\item{lambda_prior}{List or function for the error lag correlation.}

\item{rownorm}{Default to TRUE.}
}
\description{
IV - A function to perform the Kelejian and Prucha (SARAR) Model
}
