\name{netreg-package}
\alias{netreg-package}
\alias{netreg}
\docType{package}
\title{
NetReg
}
\description{
Fit graph-regularized regression models using graph background knowledge.
Graph-regularization is a novel method of penalization that uses inherent group information of coefficients and shrinks them pairwise.
}
\details{
\tabular{ll}{
Package: \tab netreg\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-05-15\cr
License: \tab none\cr
}
}
\author{
Simon Dirmeier

Maintainer: Simon Dirmeier <simon.dirmeier@gmx.de>
}
\references{
Friedman J., Hastie T., Hoefling H. and Tibshirani R. (2007)
 Pathwise coordinate optimization. \cr
 \emph{The Annals of Applied Statistics}\cr \cr
 Friedman J., Hastie T. and Tibshirani R. (2010)
 Regularization Paths for Generalized Linear Models via Coordinate Descent. \cr
 \emph{Journal of statistical software}\cr \cr
 Fu W. J. (1998)
 Penalized Regression: The Bridge Versus the Lasso. \cr
 \emph{Journal of computational and graphical statistics}\cr \cr
 Cheng W. and Wang W. (2014)
 Graph-regularized dual Lasso for robust eQTL mapping. \cr
 \emph{Bioinformatics}\cr \cr
}
\keyword{package regularization regression}
