% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_reg.R
\name{netreg}
\alias{netreg}
\alias{network_reg}
\title{Network Regularization Model Search}
\usage{
network_reg(data                       = NULL,
            out                        = NULL,
            sep                        = NULL,
            header                     = TRUE,
            ar                         = TRUE,
            plot                       = TRUE,
            subgroup                   = FALSE,
            sub_feature                = 'lag & contemp',       
            sub_method                 = 'walktrap',
            confirm_subgroup           = NULL,
            paths                      = NULL,
            conv_vars                  = NULL,
            conv_length                = 16,
            conv_interval              = 1,
            mult_vars                  = NULL,
            mean_center_mult           = FALSE,
            standardize                = FALSE,
            groupcutoff                = .75,
            subcutoff                  = .5,
            diagnos                    = FALSE,
            alpha                      = .5,
            penalties                  = NULL,
            test_penalties             = FALSE,
            exogenous                  = NULL,
            lag_exogenous              = FALSE,
            interact_exogenous         = NULL,
            interact_with_exogenous    = NULL,
            predict_with_interactions  = NULL)
}
\arguments{
\item{data}{The path to the directory where individual data files are located,
or the name of the list containing individual data. Each file or matrix within the list
must contain a single matrix containing the a T (time) by p (number of variables) matrix,
where the rows represent time and columns represent individual variables. Individuals may
have different numbers of observations (T), but must have the same number of variables (p).}

\item{out}{(Optional) The path to directory where results will be stored. If specified,
a copy of output data will be saved into the directory. If the specified directory does
not exist, it will be created.}

\item{sep}{Spacing scheme for input files. 
'' indicates space-separated; ',' indicates comma separated; '/t' indicates tab-separated
Only necessary when reading in files from physical directory.}

\item{header}{(Logical) Indicate TRUE if variable names incluced in input file, FALSE otherwise.
Only necessary when reading in files from physical directory.}

\item{ar}{(Logical) If TRUE, begin model search with all autoregressive pathways estimated
with no shrinkage (i.e., penalty = 0).}

\item{conv_vars}{Vector of variable names to be convolved via smoothed Finite Impulse 
Response (sFIR). Note, conv_vars are not not automatically considered exogenous variables.
To treat conv_vars as exogenous use the exogenous argument. Variables listed in conv_vars 
must be binary variables. If there is missing data in the endogenous variables their values 
will be imputed for the convolution operation only. Defaults to NULL. ### If there are multiple 
variables listed in conv_vars they are not used in the convolution of additional conv_vars.## 
You can't do lagged variables.}

\item{conv_length}{Expected response length in seconds. For functional MRI BOLD, 16 seconds (default) is typical
for the hemodynamic response function.}

\item{conv_interval}{Interval between data acquisition. Currently must be a constant. For 
fMRI studies, this is the repetition time. Defaults to 1.}

\item{alpha}{Elastic-net parameter for the regularization approach. Values close to 0 mimic 
the ridge penalty, which tends to shrink correlated parameters towards one another. Values 
close to 1 mimic the lasso penalty, which tends to select one parameter and shrink
the others. The default value (alpha=.5) balances these two considerations, and tends to select
groups of correlated parameters and shrink other groups towards zero.}

\item{penalties}{(Optional) A matrix of user-provided penalties to initialize group-model search. 
Should contain a column for all variables (including lagged versions and interactions) that will 
be included in the model search. Values of 1 (the default) will initilize a variable to be 
normally considered in the regularization, values of 0 will initilize a variable to be estimated
(i.e., no shrinkage), and values of Inf will exlcude variables from the model.

 @param test_penalties (Optional, Logical) Optional argument to output a sample penalty matrix
 based on function parameters. Helpful for specifying a matrix to use in the penalties arguement.
 Function will exit gracefully before running anything if test_penalties = TRUE.}

\item{exogenous}{(Optional) A list of user-specified variables to consider as exogenous
(e.g., cannot be predicted) in the model search procedure. If variable names are supplied,
variables should be referred to by name. If not, then variables should be referenced by
the pattern 'V#', where # represents the column number in the original data file (e.g., 'V5').}

\item{lag_exogenous}{(Optional, Logical) If TRUE, a lagged version of the exogenous variable 
will be created. If set to TRUE, but exogenous variables are not indicated in the arguement
above, the function will not run properly.}

\item{interact_exogenous}{(Optional) Select which exogenous variables are used to create
interaction terms with other predictors (excluding other exogenous variables). If 
lag_exogenous = TRUE, then can specify both contemporaneous and lagged versions of exogenous
variables (e.g., c('V5', 'V4Lag')). Specify as 'all' to create interactions using all exogenous 
variables (contemporaneous or lagged) present in the data. Interactions between exogenous variables 
should be created manually and included as separate columns in the input file or list.}

\item{interact_with_exogenous}{(Optional) Select which endogenous variables are combined
with the exogenous variables specified with interact_exogenous to create interaction variables
(e.g., c('V1','V2','V2Lag')). Specify as 'all' to create interaction using all endogenous variables
in the data.}

\item{predict_with_interactions}{(Optional) Select which endogenous variables should be predicted
by interaction variables. This option cannot be used if interact_exogenous or interact_with_exogenous
are NULL.}

\item{group_cutoff}{Cutoff value for inclusion of a given path at the group-level.
For instance, group_cutoff = .75 indicates that a path needs to be estimated for 75% of
individuals to be included as a group-level path.}
}
\description{
This function utilizes regression with regularization to build models for individuals 
consisting of individual and group-level paths.
}
