% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.gaussian.edgenet}
\alias{predict.gaussian.edgenet}
\title{Predict method for gaussian edgenet fits}
\usage{
\method{predict}{gaussian.edgenet}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{a fitted object of class \emph{gaussian.edgenet}}

\item{newdata}{a new (\code{m} x \code{p})-dimensional design matrix with a 
variable number of observations \code{m}, but a constant number 
of co-variables \code{p}}

\item{...}{further arguments}
}
\value{
A (\code{m} x \code{q})-dimensional matrix
}
\description{
Predicts the estimated Y.hat values for a newdata design matrix X 
similar to the other predict methods, e.g. from glm and glmnet
}
\examples{
X <- matrix(rnorm(100*10),100,10)
Y <- matrix(rnorm(100*10),100,10)
G.X <- matrix(rpois(10*10,1),10)
G.X <- t(G.X) + G.X
diag(G.X) <- 0
fit <- edgenet(X=X, Y=Y, G.X=G.X, family="gaussian")
pred <- predict(fit, X)
}
\author{
Simon Dirmeier, \email{netreg@simon-dirmeier.net}
}

