% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.edgenet}
\alias{print.edgenet}
\title{Print method for edgenet objects}
\usage{
print.edgenet(x, ...)
}
\arguments{
\item{x}{an object of class \emph{netreg.graph}}

\item{...}{further arguments}
}
\description{
Print method for objects of class \emph{edgenet} for easy vizualization of fitted edge-regularized models
}
\author{
Simon Dirmeier, \email{netreg@simon-dirmeier.net}
}

