% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nam.R
\name{print.nam}
\alias{print.nam}
\title{Print method for nam objects}
\usage{
print.nam(x, ...)
}
\arguments{
\item{x}{An object of class "nam"}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for nam objects
}
