% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_norm.R
\name{row_norm}
\alias{row_norm}
\title{row_norm -  A function where you input a matrix or dgCMatrix and you receive
the row-normalized version of it (or the reverse).}
\usage{
row_norm(Network, reverse = FALSE, sparse = FALSE)
}
\arguments{
\item{reverse}{If TRUE, converts a row-normalized matrix into one that isn't
row normalized.}

\item{sparse}{Only applies if reverse=TRUE. Conditional if you want the
output to be a sparse matrix. If reverse=FALSE, the output will maintain the
class of the input.}

\item{Netowrk}{The adjacency object to row normalize (or un-row normalize).}
}
\value{
A matrix or dgCMatrix that is either row normalized or not depending
on reverse:
}
\description{
row_norm -  A function where you input a matrix or dgCMatrix and you receive
the row-normalized version of it (or the reverse).
}
