% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndm.R
\name{summary.disturbances}
\alias{summary.disturbances}
\title{Summary method for ndm objects}
\usage{
summary.disturbances(object, probs = c(0.025, 0.975), ...)
}
\arguments{
\item{object}{An object of secondary class "disturbances"}

\item{probs}{A numeric vector of probabilities for the interval estimates.}

\item{...}{Additional arguments}
}
\description{
Summary method for ndm objects
}
