% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nem.R
\name{summary.effects}
\alias{summary.effects}
\title{Summary method for nem objects}
\usage{
summary.effects(object, probs = c(0.025, 0.975), ...)
}
\arguments{
\item{object}{An object of secondary class "effects"}

\item{probs}{A numeric vector of probabilities for the interval estimates.
Default is c(0.025, 0.975).}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for nem objects
}
