% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr.R
\name{summary.nmr}
\alias{summary.nmr}
\title{Summary method for nmr objects}
\usage{
summary.nmr(object, probs = c(0.025, 0.975), ...)
}
\arguments{
\item{object}{An object of secondary class "nmr"}

\item{probs}{A numeric vector of probabilities for the interval estimates.}

\item{...}{Additional arguments}
}
\description{
Summary method for nmr objects
}
