% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeast}
\alias{yeast}
\title{A sample yeast data set for regression}
\format{
A \code{list} containing three matrices
}
\usage{
data(yeast)
}
\description{
The yeast data set is a \code{list} containing three matrices that
can be used as an example for using \code{netReg}. The data have been taken
from the references listed below.
}
\details{
\itemize{
    \item \code{X}  (112 x 500)-dimensional binary matrix of 500
      genetic markers for 112 yeast samples
    \item \code{Y}  (112 x 231)-dimensional double matrix of 231
      gene expression values for 112 yeast samples
    \item \code{GY} (231 x 231)-dimensional adjaceny matrix representing
      protein-protein interactions for 231 yeast genes
 }
}
\references{
Brem, Rachel B., et al. (2005),
 Genetic interactions between polymorphisms that affect gene expression in
 yeast. \cr
 \emph{Nature} \cr \cr
 Storey, John D., Joshua M. Akey, and Leonid Kruglyak (2005),
 Multiple locus linkage analysis of genomewide expression in yeast. \cr
 \emph{PLoS Biology} \cr \cr
}
\keyword{data}
\keyword{datasets}
