% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\docType{methods}
\name{family}
\alias{family}
\alias{gaussian}
\alias{binomial}
\alias{poisson}
\alias{gamma}
\alias{beta}
\alias{bum}
\alias{inverse.gaussian}
\title{Family objects for models}
\usage{
family(object, ...)

gaussian(link = c("identity"))

binomial(link = c("logit", "probit", "log"))

poisson(link = c("log"))

gamma(link = c("inverse"))

beta(link = c("logit", "probit", "log"))

bum(link = c("logit", "probit", "log"))

inverse.gaussian(link = c("1/mu^2"))
}
\arguments{
\item{object}{a object for which the family shoulr be retured
(e.g. \code{edgenet})}

\item{...}{further arguments passed to methods}

\item{link}{name of a link function}
}
\value{
An object of class \code{netReg.family}
 \item{family }{ name of the family}
 \item{link }{ name of the link function}
 \item{linkinv }{ inverse link function}
 \item{loss }{ loss function}
}
\description{
Family objects provide a convenient way to specify the details
 of the models used by \code{netReg}.
 See also \code{\link[stats:family]{stats::family}} for more details.
}
\examples{
gaussian()
binomial("probit")$link
poisson()$linkinv
gamma()$linkinv
beta()$loss
inverse.gaussian()$loss
}
