% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nam.R
\name{nam}
\alias{nam}
\title{nam - A wrapper function for performing the different network autocorrelation
family of models}
\usage{
nam(
  formula,
  data = NULL,
  network,
  model = "effects",
  method = "normal",
  rownorm = TRUE,
  ...
)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced
to that class): a symbolic description of the model to be fitted.}

\item{data}{an optional data frame, list or environment (or object coercible
by as.data.frame to a data frame) containing the variables in the model.
If not found in data, the variables are taken from environment(formula),
typically the environment from which nam is called. You can also input an
igraph object and the dataset and adjacency matrix will be extracted for
you (in which case the network argument will be ignored).}

\item{network}{An adjacency matrix describing your network. Can be of the class:
c("matrix", "dgCMatrix", "igraph", "network") or an edgelist.}

\item{model}{A character indicating the model to be run. options are
c("effects", "disturbances", "NMR", "IV", "HA"). Run ?models.nam()
for a more detailed explanation and citations on each method.}

\item{method}{A character indicating the method to use for simulation.
Options vary by model but all potential options are c("normal",
"greta", "stan", "lnam"). Run ?methods.nam() for model-specific designations.}

\item{rownorm}{A logical indicating whether the inputted adjacency matrix
should be row-normalized (TRUE). Ignored by certain models that require
normalization.}

\item{...}{Additional arguments that are method-specific. Run
?args.nam() for a more detailed explanation on the different arguments for
each method.}
}
\value{
A list with two classes, the first being "nam" and the second is the
class of the method that contains values:

\code{estimates} A list containing the estimates of the parameters
of the chosen method.

\code{A} The adjacency matrix. Will be row-normalized if rownorm=TRUE.

\code{loglik} The log of the likelihood of the model.

\code{AIC} The Akaike Information Criterion of the model.

\code{BIC} The Bayesian Information Criterion of the model.
}
\description{
nam - A wrapper function for performing the different network autocorrelation
family of models
}
