% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netReg-package.R
\docType{package}
\name{netReg-package}
\alias{netReg-package}
\title{The \emph{netReg} package}
\description{
\emph{netReg} is a package for generalized linear regression that includes
prior graphs in the models objective function.
}
\details{
\emph{netReg} uses \emph{Armadillo} and \emph{TensorFlow} for
fast matrix computations and optimization.
}
\references{
Dirmeier, Simon and Fuchs, Christiane and Mueller, Nikola S and Theis,
 Fabian J (2018),
 netReg: Network-regularized linear models for biological association
 studies. \cr
 \emph{Bioinformatics}\cr \cr
 Abadi, Martín et al. (2016),
 Tensorflow: A system for large-scale machine learning. \cr
 \emph{12th USENIX Symposium on Operating Systems Design and Implementation (OSDI 16)}\cr \cr
 Powell M.J.D. (2009),
 The BOBYQA algorithm for bound constrained optimization
  without derivatives.\cr
 \url{http://www.damtp.cam.ac.uk/user/na/NA_papers/NA2009_06.pdf}
 Eddelbuettel, Dirk and Sanderson, Conrad (2014),
 RcppArmadillo: Accelerating R with high-performance C++ linear algebra.
 \emph{Computational Statistics & Data Analysis}
}
\author{
Simon Dirmeier | \email{simon.dirmeier@gmx.de}
}
\keyword{package}
